feat:候选人关系列表

+++ forum-service/src/main/java/com/github/liuyueyi/forum/service/user/UserRelationService.java
@@ -20,7 +20,7 @@ public interface UserRelationService {
      * @param pageParam
      * @return
      */
-    List<UserFollowDTO> getUserRelationList(Integer userId, PageParam pageParam);
+    List<UserFollowDTO> getUserRelationList(Long userId, PageParam pageParam);
 
     /**
      * 删除用户关系
